broke scanner into sub module
